refactor: 모델 구조 변경  - domain.model에 있는 모델들이 common model  - presentation이나 data에서는 각자 dto로 model을 만들어 쓰고 domain model로 바꾸어 로직을 수행한다.

+++ app/src/main/java/com/won983212/mongle/domain/repository/FavoriteRepository.kt
@@ -1,6 +1,6 @@
 package com.won983212.mongle.domain.repository
 
-import com.won983212.mongle.data.source.local.model.Favorite
+import com.won983212.mongle.domain.model.Favorite
 import java.time.YearMonth
 
 interface FavoriteRepository {
@@ -10,5 +10,5 @@ interface FavoriteRepository {
 
     suspend fun insert(favorite: Favorite)
 
-    suspend fun deleteById(id: Int)
+    suspend fun deleteById(id: Long)
 }
