+++ app/src/main/java/com/won983212/mongle/domain/usecase/favorite/DeleteFavoriteUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class DeleteFavoriteUseCase @Inject constructor(
     private val favoriteRepository: FavoriteRepository
 ) {
-    suspend operator fun invoke(id: Int) {
+    suspend operator fun invoke(id: Long) {
         return favoriteRepository.deleteById(id)
     }
 }
