+++ app/src/main/java/com/won983212/mongle/presentation/view/favorite/FavoriteListAdapter.kt
@@ -3,7 +3,8 @@ package com.won983212.mongle.presentation.view.favorite
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.won983212.mongle.data.source.local.model.Favorite
+import com.won983212.mongle.data.source.local.entity.FavoriteEntity
+import com.won983212.mongle.domain.model.Favorite
 import com.won983212.mongle.presentation.base.AdapterDiffCallback
 
 class FavoriteListAdapter(
@@ -26,10 +27,10 @@ class FavoriteListAdapter(
 
     // Adapter는 getItemId, containsItem가 구현되어있어야 스스로 fragment를 recreate한다.
     override fun getItemId(position: Int): Long {
-        return differ.currentList[position].id.toLong()
+        return differ.currentList[position].id
     }
 
     override fun containsItem(itemId: Long): Boolean {
-        return differ.currentList.any { it.id == itemId.toInt() }
+        return differ.currentList.any { it.id == itemId }
     }
 }
